backend/requirements.txt
Django
djangorestframework
google-genai
django-cors-headers
python-dotenv
psycopg2-binary
requests
gunicorn
whitenoise  